modernise
- update travis

- make python package

- avoid import in init

- fix versioneer

- fix cat syntax
