+++ docs/source/api.rst
@@ -3,4 +3,4 @@ API Reference
 
 .. autoclass:: intake_postgres.Plugin
 
-.. autoclass:: intake_postgres.PostgresSource
+.. autoclass:: intake_postgres.intake_postgres.PostgresSource
