+++ setup.cfg
@@ -1,7 +1,6 @@
 [versioneer]
 VCS = git
 style = pep440
-versionfile_source = ./_version.py
-versionfile_build = ../../_version.py
+versionfile_source = intake_postgres/_version.py
+versionfile_build = intake_postgres/_version.py
 tag_prefix =
-parentdir_prefix = intake-postgres-
